requirements.txt
flask==3.0.3
flask-cors==4.0.1
llama-cpp-python==0.3.2 
numpy>=1.24
jinja2>=3.1
diskcache>=5.6
google-genai>=0.3.0


#Visual studio desktop development with C++ 


For Remastered 

google-genai>=0.3.0
flask==3.0.3
flask-cors==4.0.1